+++ webpack.config.js
@@ -139,10 +139,13 @@ export default [
       fallback: {
         fs: false,
         path: false
-      },
+      }
+      /*
+      TODO: uncomment this when cheerioPolyfill is done
       alias: {
         'cheerio/slim': path.resolve(__dirname, 'cheerioPolyfill.js')
       }
+      */
     }
   },
 
@@ -185,10 +188,13 @@ export default [
       fallback: {
         fs: false,
         path: false
-      },
+      }
+      /*
+      TODO: uncomment this when cheerioPolyfill is done
       alias: {
         'cheerio/slim': path.resolve(__dirname, 'cheerioPolyfill.js')
       }
+      */
     }
   },
 
@@ -231,10 +237,13 @@ export default [
       fallback: {
         fs: false,
         path: false
-      },
+      }
+      /*
+      TODO: uncomment this when cheerioPolyfill is done
       alias: {
         'cheerio/slim': path.resolve(__dirname, 'cheerioPolyfill.js')
       }
+      */
     }
   },
 
@@ -280,10 +289,13 @@ export default [
       fallback: {
         fs: false,
         path: false
-      },
+      }
+      /*
+      TODO: uncomment this when cheerioPolyfill is done
       alias: {
         'cheerio/slim': path.resolve(__dirname, 'cheerioPolyfill.js')
       }
+      */
     }
   },
 
@@ -326,10 +338,13 @@ export default [
       fallback: {
         fs: false,
         path: false
-      },
+      }
+      /*
+      TODO: uncomment this when cheerioPolyfill is done
       alias: {
         'cheerio/slim': path.resolve(__dirname, 'cheerioPolyfill.js')
       }
+      */
     }
   },
 
@@ -372,10 +387,13 @@ export default [
       fallback: {
         fs: false,
         path: false
-      },
+      }
+      /*
+      TODO: uncomment this when cheerioPolyfill is done
       alias: {
         'cheerio/slim': path.resolve(__dirname, 'cheerioPolyfill.js')
       }
+      */
     }
   }